+++ packages/host-papp/src/sso/sessionManager/scale/signingResponse.ts
@@ -1,15 +1,15 @@
 import type { CodecType } from 'scale-ts';
 import { Bytes, Option, Result, Struct, str } from 'scale-ts';
 
-export type SignPayloadResponseData = CodecType<typeof SignPayloadResponseDataCodec>;
-export const SignPayloadResponseDataCodec = Struct({
+export type SigningPayloadResponseData = CodecType<typeof SigningPayloadResponseDataCodec>;
+export const SigningPayloadResponseDataCodec = Struct({
   signature: Bytes(),
   signedTransaction: Option(Bytes()),
 });
 
-export type SignPayloadResponse = CodecType<typeof SigningResponseCodec>;
+export type SigningPayloadResponse = CodecType<typeof SigningResponseCodec>;
 export const SigningResponseCodec = Struct({
   // referencing to RemoteMessage.messageId
   respondingTo: str,
-  payload: Result(SignPayloadResponseDataCodec, str),
+  payload: Result(SigningPayloadResponseDataCodec, str),
 });