+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+## 0.6.5 (2026-02-27)
+
+### 🚀 Features
+
+- Support updated statement store api ([#33](https://github.com/Polkadot-Community-Foundation/triangle-js-sdks/pull/33))
+
+
+## 0.6.4 (2026-02-27)
+
+### 🚀 Features
+
+- OptionBool codec
+
+### 🩹 Fixes
+
+- Small custom renderer api changes ([ce7961c](https://github.com/Polkadot-Community-Foundation/triangle-js-sdks/commit/ce7961c))
+
+### ❤️ Thank You
+
+- Sergey Zhuravlev @johnthecat
+
+## 0.6.3 (2026-02-27)
+
+### 🩹 Fixes
+
+- Fix codec for custom renderer's Button and Text 
+- Pass message id to chat custom renderer
+- Remove Error throw inside PAPI adapter if chain is not supported.
+
+### ❤️ Thank You
+
+- Sergey Zhuravlev @johnthecat
+
 ## 0.6.1 (2026-02-20)
 
 ### 🩹 Fixes